reorganized test folder
