+++ test/front_end/course_pairings.spec.js
@@ -1,7 +1,7 @@
 // This is the mocha testing file for settings component
 const chai = require("chai");
 const expect = chai.expect;
-import CoursePairings from "./../src/views/course_pairings.vue";
+import CoursePairings from "./../../src/views/course_pairings.vue";
 import { shallowMount, createLocalVue } from "@vue/test-utils";
 import VueRouter from "vue-router";
