--- test/test.spec.js
@@ -1,9 +1,0 @@
-const chai = require('chai')
-
-const expect = chai.expect
-
-describe('first test', () => {
-  it('is true', () => {
-    expect(true).to.equal(true)
-  })
-})
